+++ packages/shared/sdk-client/src/LDClientImpl.test.ts
@@ -3,6 +3,7 @@ import {
   basicPlatform,
   hasher,
   logger,
+  MockStreamingProcessor,
   setupMockStreamingProcessor,
 } from '@launchdarkly/private-js-mocks';
 
@@ -12,13 +13,13 @@ import { Flags } from './types';
 
 jest.mock('@launchdarkly/js-sdk-common', () => {
   const actual = jest.requireActual('@launchdarkly/js-sdk-common');
-  const { MockStreamingProcessor } = jest.requireActual('@launchdarkly/private-js-mocks');
+  const actualMock = jest.requireActual('@launchdarkly/private-js-mocks');
   return {
     ...actual,
     ...{
       internal: {
         ...actual.internal,
-        StreamingProcessor: MockStreamingProcessor,
+        StreamingProcessor: actualMock.MockStreamingProcessor,
       },
     },
   };
@@ -147,7 +148,7 @@ describe('sdk-client object', () => {
 
   test('identify success', async () => {
     defaultPutResponse['dev-test-flag'].value = false;
-    const carContext: LDContext = { kind: 'car', key: 'mazda-cx7' };
+    const carContext: LDContext = { kind: 'car', key: 'test-car' };
 
     await ldc.identify(carContext);
     const c = ldc.getContext();
@@ -155,17 +156,45 @@ describe('sdk-client object', () => {
 
     expect(c).toEqual({
       kind: 'multi',
-      car: { key: 'mazda-cx7' },
+      car: { key: 'test-car' },
       ...autoEnv,
     });
     expect(all).toMatchObject({
       'dev-test-flag': false,
     });
+    expect(MockStreamingProcessor).toHaveBeenCalledWith(
+      expect.anything(),
+      expect.anything(),
+      '/stream/path',
+      expect.anything(),
+      undefined,
+      expect.anything(),
+    );
+  });
+
+  test('identify success withReasons', async () => {
+    const carContext: LDContext = { kind: 'car', key: 'test-car' };
+    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, basicPlatform, {
+      logger,
+      sendEvents: false,
+      withReasons: true,
+    });
+
+    await ldc.identify(carContext);
+
+    expect(MockStreamingProcessor).toHaveBeenCalledWith(
+      expect.anything(),
+      expect.anything(),
+      '/stream/path?withReasons=true',
+      expect.anything(),
+      undefined,
+      expect.anything(),
+    );
   });
 
   test('identify success without auto env', async () => {
     defaultPutResponse['dev-test-flag'].value = false;
-    const carContext: LDContext = { kind: 'car', key: 'mazda-cx7' };
+    const carContext: LDContext = { kind: 'car', key: 'test-car' };
     ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Disabled, basicPlatform, {
       logger,
       sendEvents: false,
@@ -209,7 +238,7 @@ describe('sdk-client object', () => {
 
   test('identify error stream error', async () => {
     setupMockStreamingProcessor(true);
-    const carContext: LDContext = { kind: 'car', key: 'mazda-3' };
+    const carContext: LDContext = { kind: 'car', key: 'test-car' };
 
     await expect(ldc.identify(carContext)).rejects.toMatchObject({
       code: 401,
@@ -225,10 +254,10 @@ describe('sdk-client object', () => {
 
     await ldc.identify(context);
 
-    const carContext1: LDContext = { kind: 'car', key: 'mazda-cx' };
+    const carContext1: LDContext = { kind: 'car', key: 'test-car' };
     await ldc.identify(carContext1);
 
-    const carContext2: LDContext = { kind: 'car', key: 'subaru-forrester' };
+    const carContext2: LDContext = { kind: 'car', key: 'test-car-2' };
     await ldc.identify(carContext2);
 
     expect(emitter.listenerCount('change')).toEqual(1);